feat(arc-1210): move renderTempAccesslabel to VisitorSpaceSearchPage
