+++ src/modules/shared/components/MediaCardList/MediaCardList.types.ts
@@ -13,4 +13,5 @@ export interface MediaCardListProps
 	buttons?: (item: MediaCardProps) => ReactNode;
 	actions?: (item: MediaCardProps) => ReactNode;
 	wrapper?: (card: ReactNode, item: MediaCardProps) => ReactNode;
+	tempAccessLabel?: ReactNode;
 }
